# apps/web/src/lib/api/openapi.json
{
  "components": {
    "schemas": {
...
      "CloseIn": {
        "properties": {
          "note": {
            "anyOf": [
              {
                "maxLength": 5000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "resolution": {
            "enum": [
              "resolved",
              "dismissed"
            ],
            "title": "Resolution",
            "type": "string"
          }
        },
        "required": [
          "resolution"
        ],
        "title": "CloseIn",
        "type": "object"
      },
...
      "ExceptionCode": {
        "enum": [
          "UNREADABLE_DOCUMENT",
          "LOW_CONFIDENCE_FIELD",
          "LINE_MATH_MISMATCH",
          "TOTAL_MISMATCH",
          "TAX_MISMATCH",
          "INVALID_DATE",
          "UNKNOWN_SUPPLIER",
          "BANK_DETAILS_CHANGED",
          "POSSIBLE_DUPLICATE",
          "NO_PO",
          "PO_NOT_FOUND",
          "PRICE_VARIANCE",
          "QTY_VARIANCE",
          "RECEIPT_MISSING",
          "QTY_NOT_RECEIVED",
          "PO_OVERBILLED",
          "CURRENCY_MISMATCH",
          "ABOVE_APPROVAL_LIMIT"
        ],
        "title": "ExceptionCode",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "InvoiceStatus": {
        "enum": [
          "received",
          "extracting",
          "extracted",
          "checking",
          "cleared",
          "needs_review",
          "approved",
          "rejected",
          "exported",
          "failed"
        ],
        "title": "InvoiceStatus",
        "type": "string"
      },
...
      "NoteIn": {
        "properties": {
          "note": {
            "anyOf": [
              {
                "maxLength": 5000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "title": "NoteIn",
        "type": "object"
      },
...
      "QueueOut": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/QueueItem"
            },
            "title": "Items",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total"
        ],
        "title": "QueueOut",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Invoice Intake Agent API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/exceptions/{exception_id}/close": {
      "post": {
        "description": "Resolve or dismiss an exception. A block exception needs a note.",
        "operationId": "close_exception_exceptions__exception_id__close_post",
        "parameters": [
          {
            "in": "path",
            "name": "exception_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Exception Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CloseIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Close Exception",
        "tags": [
          "invoices"
        ]
      }
    },
...
    "/invoices": {
      "get": {
        "description": "Invoices that need a person: worst exception first, then oldest. `status` repeats\n(default: needs_review and failed); `code` keeps invoices with that exception still open.",
        "operationId": "queue_invoices_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "$ref": "#/components/schemas/InvoiceStatus"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "code",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ExceptionCode"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Code"
            }
          },
          {
            "in": "query",
            "name": "supplier_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Supplier Id"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueueOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Queue",
        "tags": [
          "invoices"
        ]
      }
    },
    "/invoices/{invoice_id}": {
      "get": {
        "operationId": "invoice_detail_invoices__invoice_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "invoice_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Invoice Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Invoice Detail",
        "tags": [
          "invoices"
        ]
      }
    },
    "/invoices/{invoice_id}/approve": {
      "post": {
        "operationId": "approve_invoices__invoice_id__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "invoice_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Invoice Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NoteIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Approve",
        "tags": [
          "invoices"
        ]
      }
    },
    "/invoices/{invoice_id}/bank/reveal": {
      "post": {
        "description": "Show the account as read from the invoice. Revealing it is written to the history.",
        "operationId": "reveal_bank_invoices__invoice_id__bank_reveal_post",
        "parameters": [
          {
            "in": "path",
            "name": "invoice_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Invoice Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevealOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reveal Bank",
        "tags": [
          "invoices"
        ]
      }
    },
...
  }
}